CLN/TST: indexing/multiindex/test_getitem.py (#24741)
